Moved x.509 related utilities to `common`
